refactor(mcp): drain nonprod-environment-lease tool domain into a scoped pack (EP-8DC217EB BET-4)
Third BET-4 domain drain (BI-2B7EE073), same pattern as the merged doc-pack
(#2755) and screen-pack (#2757). Drains the 4-tool nonprod-environment-lease
domain out of the mcp-tools.ts executeTool switch into a scoped pack.

- New apps/web/lib/mcp/packs/nonprod-lease-pack.ts: list/claim/release/renew_
  nonprod_environment_lease. Definitions + handlers moved verbatim (each lazy-
  imports @/lib/nonprod/environment-lease, same as the switch cases); grants
  mirror agent-grants (read tool → work_capsule_read, mutating → work_capsule_write).
- mcp-tools.ts: 4 definitions + 4 switch cases removed, nonprodLeasePack
  registered in composeToolPacks. Net −209 LOC; module-size baseline retightened.
- Baseline shrunk 216 → 212 (the 4 lease tool names).

Behavior-preserving relocation — the existing mcp-tools-nonprod-environments.test.ts,
which drives these tools through executeTool, stays green (now routed via the
pack handler), proving identical dispatch. New nonprod-lease-pack.test.ts adds
pack-level assertions (names, provenance-free descriptions, grant mirror,
delegation).

Verification (worktree bet4-drain3, source-local): web typecheck exit 0 (8GB);
new pack test + existing nonprod-environments test 13/13; tool-pack ratchet
green (212); (full suite below).

Process-Spine-Decision: EP-8DC217EB plan §4 BET-4 — drain the nonprod-lease tool domain into a scoped pack; ratchet baseline shrunk 216→212; behavior-preserving relocation.
