+++ scripts/mcp-tool-pack-baseline.json
@@ -22,7 +22,6 @@
   "cancel_thread",
   "check_deployment_windows",
   "check_sandbox",
-  "claim_nonprod_environment_lease",
   "classify_ea_element",
   "compare_versions",
   "configure_gateway_scan",
@@ -99,7 +98,6 @@
   "list_departments",
   "list_epics",
   "list_my_capability_needs",
-  "list_nonprod_environment_leases",
   "list_opportunities",
   "list_patch_posture",
   "list_positions",
@@ -142,8 +140,6 @@
   "recover_sandbox",
   "refresh_channel_kpis",
   "register_digital_product_from_build",
-  "release_nonprod_environment_lease",
-  "renew_nonprod_environment_lease",
   "request_coworker",
   "resolve_model_selection",
   "resolve_portfolio_quality_issue",
